Add int conversions for Python 3 compatibility
I don't fully understand the problem here. I admit it. What happens
is that when running with Python 3 but without explicitly sourcing
the virtualenv, then the changed locations raise errors because the
sizes are floats instead of ints.
Maybe the underlying libraries are newer versions in the Python 3 packages
and thus expect different types than in Python 2...